Rework most examples so they don't need an entire sub-crate
