--- examples/linked_list/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "linked_list"
-version = "0.1.0"
-authors = ["Devin R <devin.ragotzy@gmail.com>"]
-edition = "2018"
-
-[workspace]
-
-[dependencies]
-genawaiter = { path = "../.." }
-
-[features]
-strict = []
